feat: complete cjsToEsm.mjs subtask
