+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,30 @@
+import * as path from 'path';
+import { release, version } from 'os';
+const { createServer: createServerHttp } = import('http');
+import('./files/c');
+import {fileURLToPath} from 'url';
+import {dirname} from 'path';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const random = Math.random();
+
+export let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = import('./files/a.json');
+} else {
+    unknownObject = import('./files/b.json');
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+export const createMyServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
